+++ DependabotHelper.sln
@@ -77,6 +77,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{EF9C21CA-0
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DependabotHelper.AppHost", "src\DependabotHelper.AppHost\DependabotHelper.AppHost.csproj", "{1E3C26BD-B8D6-4E76-9558-586657C57301}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "perf", "perf", "{5814B1E0-6694-45D4-9C46-495B865DC5A8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DependabotHelper.Benchmarks", "perf\DependabotHelper.Benchmarks\DependabotHelper.Benchmarks.csproj", "{9020A582-E59F-4D11-8F09-61A53AF62D45}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -99,6 +103,10 @@ Global
 		{1E3C26BD-B8D6-4E76-9558-586657C57301}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1E3C26BD-B8D6-4E76-9558-586657C57301}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1E3C26BD-B8D6-4E76-9558-586657C57301}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9020A582-E59F-4D11-8F09-61A53AF62D45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9020A582-E59F-4D11-8F09-61A53AF62D45}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9020A582-E59F-4D11-8F09-61A53AF62D45}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9020A582-E59F-4D11-8F09-61A53AF62D45}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -113,6 +121,7 @@ Global
 		{D3A273BE-65C1-45EA-A1F3-BBCC2E6F0282} = {FD791692-13A5-47D7-ACEB-4B0B336B359D}
 		{EF9C21CA-0F1C-4E1F-A61D-3F90F237DF90} = {2808D3E0-7B58-4A2A-95BE-8027C4ED9918}
 		{1E3C26BD-B8D6-4E76-9558-586657C57301} = {922D7DF2-F45E-4D41-9287-DC3F6A447671}
+		{9020A582-E59F-4D11-8F09-61A53AF62D45} = {5814B1E0-6694-45D4-9C46-495B865DC5A8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2F443F99-DDBC-4E35-A6A9-E790CA137404}
